Moved file-related functions from runner.ts to src/files/ or src/utils.ts
src/runner.ts is getting pretty overcrowded; it was getting hard to reason about where things belonged in it.
