+++ src/utils.ts
@@ -235,6 +235,10 @@ export function detectBufferEncoding(buffer: Buffer, length = buffer.length): En
     return "utf8";
 }
 
+export function trimSingleQuotes(str: string): string {
+    return str.replace(/^'|'$/g, "");
+}
+
 // converts Windows normalized paths (with backwards slash `\`) to paths used by TypeScript (with forward slash `/`)
 export function denormalizeWinPath(path: string): string {
     return path.replace(/\\/g, "/");
